Support for DELAY as a floating point number including less than 1 second while maintaining backward compatibiitiy with integer arguments.

+++ image.h
@@ -60,7 +60,7 @@ typedef struct {
 
 	struct {
 		bool on;
-		int delay;
+		float delay;
 	} ss;
 
 	multi_img_t multi;
